POC for jupyterhub in iframe
